gate rollback side effects behind the window pre-check, move pre-flight e2e to the config-owning spec, keep typed handle as login hint
