+++ apps/app/src/server/auth/rollback.ts
@@ -29,27 +29,37 @@ export async function revokeAtprotoGrantsBeforeUserDeletion(
  * request cookies, which the after-hook's request does not carry — the new
  * session only exists on the response — so rollback deletes directly.
  *
- * The user delete is scoped to the rollback window so this primitive
- * cannot destroy an established account even through a buggy caller; when
- * it matches nothing, every related row is left untouched. The account,
- * session, and connection deletes duplicate the schema's FK cascades on
- * purpose: with foreign-key enforcement ever off, a surviving orphan
- * account row would read as an existing account at the authorize
- * pre-flight and silently re-open the signup bypass.
+ * Every destructive step is behind the rollback-window guard, so this
+ * primitive cannot harm an established account even through a buggy
+ * caller: outside the window nothing at all is touched — not even the
+ * grant revocation, which would otherwise sever a real user's Atmosphere
+ * connection. The guarded delete inside the transaction re-checks the
+ * window as the TOCTOU backstop. The account, session, and connection
+ * deletes duplicate the schema's FK cascades on purpose: with foreign-key
+ * enforcement ever off, a surviving orphan account row would read as an
+ * existing account at the authorize pre-flight and silently re-open the
+ * signup bypass.
  */
 export async function rollbackAutoCreatedUser(userId: string): Promise<void> {
+  const cutoff = new Date(Date.now() - AUTO_SIGNUP_ROLLBACK_WINDOW_MS);
+  const justCreated = await db
+    .select({ id: user.id })
+    .from(user)
+    .where(and(eq(user.id, userId), gte(user.createdAt, cutoff)))
+    .get();
+  if (!justCreated) {
+    logError(
+      `[auth] auto-signup rollback matched no just-created user row for ${userId}; leaving every related row untouched`,
+    );
+    return;
+  }
+
   await revokeAtprotoGrantsBeforeUserDeletion(userId);
   await db.transaction(async (tx) => {
-    const cutoff = new Date(Date.now() - AUTO_SIGNUP_ROLLBACK_WINDOW_MS);
     const deleted = await tx
       .delete(user)
       .where(and(eq(user.id, userId), gte(user.createdAt, cutoff)));
-    if (deleted.rowsAffected === 0) {
-      logError(
-        `[auth] auto-signup rollback matched no just-created user row for ${userId}; leaving related rows untouched`,
-      );
-      return;
-    }
+    if (deleted.rowsAffected === 0) return;
     await tx.delete(session).where(eq(session.userId, userId));
     await tx.delete(account).where(eq(account.userId, userId));
     await tx
