+++ apps/app/tests/e2e/self-hosted/admin-signin-methods.spec.ts
@@ -11,8 +11,15 @@ import type { Page } from "@playwright/test";
  * Admin sign-in method settings for the Atmosphere provider: the toggle
  * renders alongside email/OAuth, disabling it gates the authorize endpoint
  * server-side, and lockout accounting locks the row while an admin's only
- * sign-in method is an atproto DID.
+ * sign-in method is an atproto DID. The authorize sign-up pre-flight lives
+ * here too — this file owns provider-config mutation, so every mutator of
+ * that shared instance state shares one worker.
  */
+
+// Both describes mutate global provider config and restore it in
+// afterEach; run this file's tests in order in one worker (without serial
+// mode's skip-on-failure coupling) so they cannot interleave.
+test.describe.configure({ mode: "default" });
 
 /** Seed an admin whose only account row is an atproto DID. */
 async function seedAtprotoOnlyAdmin(tursoPort: number, email: string) {
@@ -132,5 +139,137 @@ test.describe("admin sign-in method settings", () => {
     await expect(page.locator("#signin-atproto-toggle")).toHaveCount(0, {
       timeout: 30000,
     });
+  });
+});
+
+/**
+ * The authorize sign-up pre-flight: with atproto excluded from the sign-up
+ * providers, an unknown DID is rejected before any authorize URL is
+ * issued (no create-then-roll-back at the callback), while a DID with an
+ * existing account row still gets past the sign-up gate. The pre-resolved
+ * `did` body field keeps both requests hermetic — no identity resolution
+ * leaves the instance before the gate runs.
+ */
+test.describe("atmosphere authorize sign-up pre-flight", () => {
+  /** Exclude atproto from sign-up only; sign-in stays fully enabled. */
+  async function excludeAtprotoFromSignup() {
+    const client = createClient({
+      url: `http://127.0.0.1:${SELF_HOSTED_TURSO_PORT}`,
+    });
+    try {
+      await client.execute({
+        sql: `INSERT INTO serial_app_config (key, value, updated_at)
+              VALUES ('enabled-signup-providers', ?, unixepoch())
+              ON CONFLICT(key) DO UPDATE SET value = excluded.value, updated_at = excluded.updated_at`,
+        args: [JSON.stringify(["email", "oauth"])],
+      });
+    } finally {
+      client.close();
+    }
+  }
+
+  async function seedAtprotoAccount(did: string) {
+    const client = createClient({
+      url: `http://127.0.0.1:${SELF_HOSTED_TURSO_PORT}`,
+    });
+    const userId = createId();
+    const now = Math.floor(Date.now() / 1000);
+    try {
+      await client.batch([
+        {
+          sql: `INSERT INTO serial_user (id, name, email, email_verified, image, created_at, updated_at)
+                VALUES (?, 'Preflight Known', ?, 1, NULL, ?, ?)`,
+          args: [userId, `preflight-known-${userId}@example.com`, now, now],
+        },
+        {
+          sql: `INSERT INTO serial_account (id, account_id, provider_id, user_id, created_at, updated_at)
+                VALUES (?, ?, 'atproto', ?, ?, ?)`,
+          args: [createId(), did, userId, now, now],
+        },
+      ]);
+    } finally {
+      client.close();
+    }
+    return userId;
+  }
+
+  async function cleanupSeededUser(userId: string) {
+    const client = createClient({
+      url: `http://127.0.0.1:${SELF_HOSTED_TURSO_PORT}`,
+    });
+    try {
+      await client.execute({
+        sql: "DELETE FROM serial_user WHERE id = ?",
+        args: [userId],
+      });
+    } finally {
+      client.close();
+    }
+  }
+
+  async function postAuthorize(page: Page, identifier: string, did: string) {
+    return page.evaluate(
+      async (body) => {
+        const response = await fetch("/api/auth/atproto/authorize", {
+          method: "POST",
+          headers: { "Content-Type": "application/json" },
+          body: JSON.stringify(body),
+        });
+        return {
+          status: response.status,
+          body: (await response.json()) as { message?: string },
+        };
+      },
+      { identifier, did },
+    );
+  }
+
+  test.afterEach(async () => {
+    // Restore the global-setup provider state for other specs.
+    await setEnabledAuthProviders(SELF_HOSTED_TURSO_PORT, [
+      "email",
+      "oauth",
+      "atproto",
+    ]);
+  });
+
+  test("rejects an unknown DID at authorize while atproto sign-up is unavailable", async ({
+    page,
+  }) => {
+    test.setTimeout(120000);
+    await excludeAtprotoFromSignup();
+    await page.goto("/auth/sign-in");
+
+    const rejected = await postAuthorize(
+      page,
+      "stranger.test",
+      "did:plc:e2e-preflight-stranger",
+    );
+    expect(rejected.status).toBe(400);
+    expect(rejected.body.message).toContain("Sign ups are currently disabled");
+  });
+
+  test("lets a known DID past the sign-up gate while sign-ups are unavailable", async ({
+    page,
+  }) => {
+    test.setTimeout(120000);
+    const did = "did:plc:e2e-preflight-known";
+    const userId = await seedAtprotoAccount(did);
+    try {
+      await excludeAtprotoFromSignup();
+      await page.goto("/auth/sign-in");
+
+      // The gate passes; the flow then fails downstream in this PDS-less
+      // environment. The exact generic authorize failure — asserting only
+      // "not the signups-disabled message" would also pass if the sign-in
+      // gate (wrongly) rejected the request first.
+      const outcome = await postAuthorize(page, "known.test", did);
+      expect(outcome.status).toBe(400);
+      expect(outcome.body.message).toContain(
+        "Could not start Atmosphere sign in",
+      );
+    } finally {
+      await cleanupSeededUser(userId);
+    }
   });
 });
